+++ gpiozero/testing/__init__.py
@@ -1,0 +1,3 @@
+from .fixtures import no_default_factory, mock_factory, pwm
+
+__ALL__ = ['no_default_factory', 'mock_factory', 'pwm']
